requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
packaging==25.0
playwright==1.52.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.1
pytest==8.4.0
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-slugify==8.0.4
requests==2.32.4
text-unidecode==1.3
typing_extensions==4.14.0
urllib3==2.4.0
typing_extensions~=4.14.0
Faker~=37.4.0